tests/requirements.txt
docker
ipython
ipdb
pytest
python-etcd
requests
